simplification of the collection
