--- src/MiddlewareCollection/MutableMiddlewareCollection.php
@@ -1,61 +1,0 @@
-<?php
-//
-// +---------------------------------------------------------------------+
-// | CODE INC. SOURCE CODE                                               |
-// +---------------------------------------------------------------------+
-// | Copyright (c) 2018 - Code Inc. SAS - All Rights Reserved.           |
-// | Visit https://www.codeinc.fr for more information about licensing.  |
-// +---------------------------------------------------------------------+
-// | NOTICE:  All information contained herein is, and remains the       |
-// | property of Code Inc. SAS. The intellectual and technical concepts  |
-// | contained herein are proprietary to Code Inc. SAS are protected by  |
-// | trade secret or copyright law. Dissemination of this information or |
-// | reproduction of this material is strictly forbidden unless prior    |
-// | written permission is obtained from Code Inc. SAS.                  |
-// +---------------------------------------------------------------------+
-//
-// Author:   Joan Fabrégat <joan@codeinc.fr>
-// Date:     24/09/2018
-// Project:  MiddlewareDispatcher
-//
-declare(strict_types=1);
-namespace CodeInc\MiddlewareDispatcher\MiddlewareCollection;
-use CodeInc\MiddlewareDispatcher\MiddlewareDispatcherException;
-use Psr\Http\Server\MiddlewareInterface;
-
-
-/**
- * Class MutableMiddlewareCollection
- *
- * @package CodeInc\MiddlewareDispatcher\MiddlewareCollection
- * @author Joan Fabrégat <joan@codeinc.fr>
- */
-class MutableMiddlewareCollection extends AbstractMiddlewareCollection implements MutableMiddlewareCollectionInterface
-{
-    /**
-     * MiddlewareCollection constructor.
-     *
-     * @param iterable|null $middleware
-     * @throws MiddlewareDispatcherException
-     */
-    public function __construct(?iterable $middleware = null)
-    {
-        if ($middleware !== null) {
-            foreach ($middleware as $item) {
-                if (!$item instanceof MiddlewareInterface) {
-                    throw MiddlewareDispatcherException::notAMiddleware($item);
-                }
-                $this->addMiddleware($item);
-            }
-        }
-    }
-
-    /**
-     * @inheritdoc
-     * @param MiddlewareInterface $middleware
-     */
-    public function addMiddleware(MiddlewareInterface $middleware):void
-    {
-        $this->middleware[] = $middleware;
-    }
-}
